refactor(DetailViewController): extension String의 접근제어자, 위치, 함수명 변경
