+++ Diary/Controller/DetailViewController.swift
@@ -32,7 +32,7 @@ final class DetailViewController: UIViewController {
     
     private func shareChangedData() {
         guard let indexPath = indexPath else { return }
-        let data = detailTextView.text.parseData(titleWordsLimit: 20)
+        let data = detailTextView.text.seperateTitleAndBody(titleWordsLimit: 20)
         delegate?.sendData(title: data.title,
                            body: data.body,
                            indexPath: indexPath)
@@ -86,5 +86,23 @@ extension DetailViewController {
         }
         let keyboardHeight = keyboardFrame.cgRectValue.height
         detailTextViewBottomConstraint.constant = isAppearing ? keyboardHeight : .zero
+    }
+}
+
+// MARK: - extnesion: seperate detailTextView Title & Body
+
+fileprivate extension String {
+    func seperateTitleAndBody(titleWordsLimit: Int) -> (title: String, body: String) {
+        var components = components(separatedBy: "\n\n")
+        if components.count > 1 {
+            let title = components.removeFirst()
+            let body = components.filter { $0 != ""}.joined(separator: "\n\n")
+            return (title, body)
+        }
+        let limitIndex = index(startIndex, offsetBy: titleWordsLimit)
+        let title = self[startIndex..<limitIndex]
+        let body = self[limitIndex..<endIndex]
+        
+        return (title.description, body.description)
     }
 }
